# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NewsArticle {
  id: number;
  title: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  publishedAt: string;
  createdAt: string;
}

export interface NewsInput {
  /** @minLength 1 */
  title: string;
  excerpt?: string;
  /** @minLength 1 */
  content: string;
  category: string;
  imageUrl?: string;
  publishedAt?: string;
}

export interface NewsUpdate {
  title?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  imageUrl?: string;
  publishedAt?: string;
}

export interface Event {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  date: string;
  /** @nullable */
  endDate?: string | null;
  location: string;
  category?: string;
  createdAt: string;
}

export interface EventInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  date: string;
  endDate?: string;
  location: string;
  category?: string;
}

export interface EventUpdate {
  title?: string;
  description?: string;
  date?: string;
  endDate?: string;
  location?: string;
  category?: string;
}

export interface Program {
  id: number;
  name: string;
  level: string;
  description: string;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ProgramInput {
  /** @minLength 1 */
  name: string;
  level: string;
  description: string;
  duration?: string;
  imageUrl?: string;
}

export interface ProgramUpdate {
  name?: string;
  level?: string;
  description?: string;
  duration?: string;
  imageUrl?: string;
}

export interface GalleryItem {
  id: number;
  title: string;
  imageUrl: string;
  category: string;
  /** @nullable */
  takenAt?: string | null;
}

export interface GalleryInput {
  /** @minLength 1 */
  title: string;
  imageUrl: string;
  category: string;
  takenAt?: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  subject: string;
  /** @minLength 1 */
  message: string;
  phone?: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface SchoolProfile {
  id: number;
  name: string;
  tagline: string;
  description: string;
  address: string;
  phone: string;
  email: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  npsn?: string | null;
  /** @nullable */
  accreditation?: string | null;
  /** @nullable */
  foundedYear?: number | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
}

export interface SchoolProfileUpdate {
  name?: string;
  tagline?: string;
  description?: string;
  address?: string;
  phone?: string;
  email?: string;
  website?: string;
  npsn?: string;
  accreditation?: string;
  foundedYear?: number;
  instagramUrl?: string;
  facebookUrl?: string;
  youtubeUrl?: string;
  logoUrl?: string;
}

export interface SchoolStats {
  totalStudents: number;
  totalTeachers: number;
  totalPrograms: number;
  yearsEstablished: number;
  recentNewsCount: number;
  upcomingEventsCount: number;
}

export type ListNewsParams = {
limit?: number;
offset?: number;
category?: string;
};

export type ListEventsParams = {
upcoming?: boolean;
};

export type ListGalleryParams = {
category?: string;
};