# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: School Website API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: news
    description: News and announcements
  - name: events
    description: School events
  - name: programs
    description: Academic programs
  - name: gallery
    description: Photo gallery
  - name: contact
    description: Contact form
  - name: stats
    description: School statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List all news articles
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          required: false
        - in: query
          name: offset
          schema:
            type: integer
          required: false
        - in: query
          name: category
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of news articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
    post:
      operationId: createNews
      tags: [news]
      summary: Create a news article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"

  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get a news article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update a news article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete a news article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      parameters:
        - in: query
          name: upcoming
          schema:
            type: boolean
          required: false
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List academic programs
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Create a program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"

  /programs/{id}:
    patch:
      operationId: updateProgram
      tags: [programs]
      summary: Update a program
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    delete:
      operationId: deleteProgram
      tags: [programs]
      summary: Delete a program
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery photos
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Add a gallery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"

  /gallery/{id}:
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete a gallery item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /contact:
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List all contact messages (admin)
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  /contact/{id}:
    delete:
      operationId: deleteContactMessage
      tags: [contact]
      summary: Delete a contact message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /school-profile:
    get:
      operationId: getSchoolProfile
      tags: [stats]
      summary: Get school profile
      responses:
        "200":
          description: School profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolProfile"
    patch:
      operationId: updateSchoolProfile
      tags: [stats]
      summary: Update school profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolProfile"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get school statistics summary
      responses:
        "200":
          description: School stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    NewsArticle:
      type: object
      required: [id, title, content, category, publishedAt, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        publishedAt:
          type: string
        createdAt:
          type: string

    NewsInput:
      type: object
      required: [title, content, category]
      properties:
        title:
          type: string
          minLength: 1
        excerpt:
          type: string
        content:
          type: string
          minLength: 1
        category:
          type: string
        imageUrl:
          type: string
        publishedAt:
          type: string

    NewsUpdate:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        publishedAt:
          type: string

    Event:
      type: object
      required: [id, title, date, location, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: string
        category:
          type: string
        createdAt:
          type: string

    EventInput:
      type: object
      required: [title, date, location]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        date:
          type: string
        endDate:
          type: string
        location:
          type: string
        category:
          type: string

    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        endDate:
          type: string
        location:
          type: string
        category:
          type: string

    Program:
      type: object
      required: [id, name, level, description]
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: string
        description:
          type: string
        duration:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]

    ProgramInput:
      type: object
      required: [name, level, description]
      properties:
        name:
          type: string
          minLength: 1
        level:
          type: string
        description:
          type: string
        duration:
          type: string
        imageUrl:
          type: string

    ProgramUpdate:
      type: object
      properties:
        name:
          type: string
        level:
          type: string
        description:
          type: string
        duration:
          type: string
        imageUrl:
          type: string

    GalleryItem:
      type: object
      required: [id, title, imageUrl, category]
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        takenAt:
          type: ["string", "null"]

    GalleryInput:
      type: object
      required: [title, imageUrl, category]
      properties:
        title:
          type: string
          minLength: 1
        imageUrl:
          type: string
        category:
          type: string
        takenAt:
          type: string

    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        subject:
          type: string
        message:
          type: string
          minLength: 1
        phone:
          type: string

    ContactMessage:
      type: object
      required: [id, name, email, subject, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        createdAt:
          type: string

    SchoolProfile:
      type: object
      required: [id, name, tagline, description, address, phone, email]
      properties:
        id:
          type: integer
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: ["string", "null"]
        npsn:
          type: ["string", "null"]
        accreditation:
          type: ["string", "null"]
        foundedYear:
          type: ["integer", "null"]
        instagramUrl:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]

    SchoolProfileUpdate:
      type: object
      properties:
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        npsn:
          type: string
        accreditation:
          type: string
        foundedYear:
          type: integer
        instagramUrl:
          type: string
        facebookUrl:
          type: string
        youtubeUrl:
          type: string
        logoUrl:
          type: string

    SchoolStats:
      type: object
      required: [totalStudents, totalTeachers, totalPrograms, yearsEstablished, recentNewsCount, upcomingEventsCount]
      properties:
        totalStudents:
          type: integer
        totalTeachers:
          type: integer
        totalPrograms:
          type: integer
        yearsEstablished:
          type: integer
        recentNewsCount:
          type: integer
        upcomingEventsCount:
          type: integer