# lib/api-zod/src/generated/types/galleryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Website API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryInput {
  /** @minLength 1 */
  title: string;
  imageUrl: string;
  category: string;
  takenAt?: string;
}