# lib/api-zod/src/generated/types/newsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Website API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsUpdate {
  title?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  imageUrl?: string;
  publishedAt?: string;
}